tech(TTB-060): Moved Home component to individual directory
