+++ client/src/components/Home/HomeButtons.jsx
@@ -1,4 +1,4 @@
-import Button from '../components/Button/Button';
+import Button from '../Button/Button';
 
 function HomeButtons() {
     return(
